Refactor Input and Search to component
